Load settings based on environment variable.
