+++ bin/api
@@ -1,8 +1,7 @@
 #!/bin/bash
 
-pip install -q -r requirements.txt
+make install-dependencies
 
 export FLASK_APP=snuba/api.py
-export FLASK_DEBUG=1
 
 flask run
