+++ snuba/settings_test.py
@@ -1,0 +1,4 @@
+from snuba.settings_base import *  # NOQA
+
+TESTING = True
+CLICKHOUSE_TABLE = 'test'
